fix data length type (#38)

+++ DbcParserLib/Parsers/MessageLineParser.cs
@@ -20,7 +20,7 @@ namespace DbcParserLib.Parsers
                 var msg = new Message()
                 {
                     Name = match.Groups[2].Value,
-                    DLC = byte.Parse(match.Groups[3].Value, CultureInfo.InvariantCulture),
+                    DLC = ushort.Parse(match.Groups[3].Value, CultureInfo.InvariantCulture),
                     Transmitter = match.Groups[4].Value,
                     ID = uint.Parse(match.Groups[1].Value, CultureInfo.InvariantCulture)
                 };
